crewai/debate/pyproject.toml
[project]
name = "debate"
version = "0.1.0"
description = "debate using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.14.3",
    "python-dotenv>=1.0",
]

[project.scripts]
debate = "debate.main:run"
run_crew = "debate.main:run"
train = "debate.main:train"
replay = "debate.main:replay"
test = "debate.main:test"
run_with_trigger = "debate.main:run_with_trigger"

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.crewai]
type = "crew"
